Add filter-list example into gallery
